Start writing a better scanner function
The problem we are running into with decoding larger image is almost entirely an issue with libdeflate not automatically separating IDAT chunks. There might be a good reason for it but it's annoying and it means I have to write a better scanner function to deal with it.

The way a lot of this file handlign is getting done is really stupid as well. We could just read the file instead of copying all the file data and doign stupid math on it. So thats in the works as well.